Remove folder typedata in-memory cache
Summary:
This caused the syncroton Ping command to miss new task folders.
Because the Ping command is long running we use:

    rcube::get_instance()->get_storage()->clear_cache('mailboxes', true);

To clear the cache before listing the folders. However, this doesn't
clear the in-memory cache in kolab_storage.

In list_folders we then ended up in the branch returning the folder list directly from
that outdated in-memory cache and thus missed the update.

Since we already cache the folder list and metadata it doesn't seem necessary to
cache it again. Removing it thus gives us a single place to invalidate it.

Differential Revision: https://git.kolab.org/D2143